requirements_prod.txt
# Production-specific dependencies for Google App Engine deployment

# Core Flask application (from main requirements.txt)
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Login>=0.5.0
Flask-WTF>=1.0.0
Flask-Mail>=0.9.1
Flask-Migrate>=3.0.0
Werkzeug==2.0.3
python-dotenv>=0.19.0
email-validator>=1.1.0
WTForms>=3.0.0
WTForms-SQLAlchemy>=0.3.0
Pillow>=9.0.0
requests==2.32.3
icalendar==6.1.3
recurring-ical-events==3.6.1
python-dateutil>=2.8.2
pytz>=2021.1
beautifulsoup4

# Production web server
gunicorn>=20.1.0

# Google Cloud dependencies
google-cloud-logging>=3.0.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage>=2.5.0
google-cloud-sql-connector>=1.0.0

# Database drivers
psycopg2-binary>=2.9.0
pg8000>=1.29.0  # Alternative PostgreSQL driver for Cloud SQL

# Caching and sessions (for Cloud Memorystore)
redis>=4.0.0
Flask-Caching>=2.0.0
Flask-Session>=0.4.0

# Monitoring and error tracking
sentry-sdk[flask]>=1.15.0
prometheus-client>=0.15.0

# Security and validation
bleach>=6.0.0
marshmallow>=3.19.0
Flask-Limiter>=3.0.0

# Communication services
twilio>=7.0.0
phonenumbers>=8.12.0

# File handling
python-magic>=0.4.24
boto3>=1.18.0  # For S3 compatibility if needed

# Performance monitoring
psutil>=5.9.0

# Static file handling
whitenoise[brotli]>=6.0.0